Update: extract and reuse useDisableBodyScroll
